Настраивает webpack. Разбивает код на модули
